Remove gamedig
GameDig appears to be an abandoned project, and is not updating
frequently enough to be considered a viable tool for use in this
project.

Additionally, Gamedig has at least 4 security vulnerabilities present
in its own dependencies that have gone unfixed for over a year.

Regex DoS Vuln in hawk@1.1.1 (dependency)
https://snyk.io/vuln/npm:hawk:20160119

2 Regex DoS Vulns in moment@2.29.0 (dependency)
https://snyk.io/vuln/npm:moment:20160126
https://snyk.io/vuln/npm:moment:20161019

Remote Memory Exposure Vuln in request@2.51.0 (dependency)
https://snyk.io/vuln/npm:request:20160119

For these reasons I have removed Gamedig and query support from the
daemon until a suitable replacement can be found and implemented.