Fixes a strange boot/rebuild race condition
This bug was probably around for a while in previous versions but went
undetected due to the methods employed to boot servers and handle the
rebuild process. It is also entirely possible it was introduced with
some of the 0.4.0 codebase.

Either way, this commit fixes a race condition that would occur if you
sent multiple start events to the websocket, as well as a daemon bug
when rebuilding. Previously it was possible for the daemon to loose
track of the server and run multiple socket streams, which only served
to confuse the browser even more.