Add support for limiting disk usage.
Checks disk space on startup, disallows if server is violating. Also
checks every minute while the server is running and will stop the
server if it starts violating the limit.

https://s3.kelp.in/CpHzC.png

This implementation does not stop a user from maliciously uploading a
large quantity of files. That would need to be controlled by a
different process. In a future iteration I plan to have the daemon
constantly checking, and if the disk use goes over an upper threshold
(e.g. 10% or 200MB) the server will be automatically suspend to halt
abuse.