Fixes files being written out of order
Not really sure why this is not happening to other users, but may be because I use mergerfs as my main server storage.

In short: I noticed that quite alot of files had problems when uploading them via the SFTP server. After comparing a include file, I noticed that a part of the file got written before it was supposed to. After hours of tests I realized that `offset` was being ignored when writing to disk.